requirements.txt
nemo_toolkit
fiddle
python-dotenv
nvidia-ml-py
fastapi
uvicorn
python-multipart
pydantic
torch
numpy
hydra-core
lightning
lhotse
einops
transformers
sentencepiece
pandas
jiwer
librosa
pyannote-audio
webdataset
datasets
editdistance
IPython
pydub
ffmpeg-python
soundfile
cuda-python
